Updated formatting and README
Updated the formatting in the `main.rs` files to be a little more consistent, including checking the `cargo clippy` suggestions.

Also updated the README because the `zip` command seems to have changed? not it zips the entire path as opposed to just the files, so use a subshell to `cd` to the correct directory and perform `zip` locally.